requirements.txt
python-telegram-bot==21.6
openpyxl==3.1.5
fpdf2==2.7.9
